Refactor settings dialog to simplify enable/disable logic (#7428)
* Move screen config files to new GUI lib

* Fixed bugs in config dialog related to service checkbox

* Test status tip

* Revert "Test status tip"

This reverts commit d11ad5bc22dab7edcf7ef0b3baada2fce240398a.

* Explain why service disabled

* Corrected client error message (written by non-native English writer) to make the message easier to understand for native English speakers

* Clearer logic

* Improve readability

* Coverage for 'add client' and client error messages

* Coverage for `Screen` class

* Clarify test names

* Use `const QString &` for backward compatibility

* Fixed lint warnings

* Prevent return paths after qFatal

* Use auto and abort to solve lint warnings

* Remove unused fake error

* Make operator hidden friend

* Remove redundant access specifier

* Convert int explicitly

* Swap && logic for linter

* Update ChangeLog

+++ src/lib/gui/config/ScreenList.cpp
@@ -90,7 +90,8 @@ ScreenList::ScreenList(int width) : QList<Screen>(), m_width(width) {}
 
 void ScreenList::addScreenByPriority(const Screen &newScreen) {
   int serverIndex = getServerIndex(*this);
-  auto indexes = getNeighborsIndexes(serverIndex, m_width, size());
+  auto indexes =
+      getNeighborsIndexes(serverIndex, m_width, static_cast<int>(size()));
 
   bool isAdded = false;
   for (const auto &index : indexes) {
